Move utility functions into separate file
